+++ src/Discovery/MessageBusCompilerListener.php
@@ -9,7 +9,6 @@ use Spiral\Core\Attribute\Singleton;
 use Spiral\Tokenizer\Attribute\TargetAttribute;
 use Spiral\Tokenizer\TokenizationListenerInterface;
 use Wolfcharaa\MessageBus\Attribute\CommandHandler;
-use Wolfcharaa\MessageBus\Attribute\DomainHandler;
 use Wolfcharaa\MessageBus\Attribute\EventSubscriber;
 use Wolfcharaa\MessageBus\Attribute\MessageAlias;
 use Wolfcharaa\MessageBus\Attribute\QueryHandler;
@@ -21,7 +20,6 @@ use Wolfcharaa\MessageBus\Spiral\Application\Config\MessageBusConfig;
 #[Singleton]
 #[TargetAttribute(CommandHandler::class)]
 #[TargetAttribute(QueryHandler::class)]
-#[TargetAttribute(DomainHandler::class)]
 #[TargetAttribute(EventSubscriber::class)]
 #[TargetAttribute(MessageAlias::class)]
 final class MessageBusCompilerListener implements TokenizationListenerInterface